exp: explicitly format in csc to avoid warnings
Computing sp.linalg.splu of a csr matrix results in a  SparseEfficiency warning:

/home/joseph/phd/HyperHDG/experiments/fiber_network_elastic.py:126: SparseEfficiencyWarning: splu converted its input to CSC format
  self.splu_precond_lhs = sp.linalg.splu(precond_lhs)

Now explicitly use csc for A and the precond_lhs.
self.coarse_basis_int is csc and hence it's transpose is csr,
and a csr @ csc product becomes csr (scipy internal choice).
Therefore explicit cast to csc is neccessary.